Render with Vue
