+++ src/frontend/components/SearchField.vue
@@ -1,0 +1,14 @@
+<template lang="html">
+  <div id="inputWrapper" class="col-xs-12">
+    <input id="input" type="text" class="col-xs-12" />
+  </div>
+</template>
+
+<script>
+module.exports = {
+
+};
+</script>
+
+<style lang="css">
+</style>
